Move cli deps in cli and remote renderer/deploy.go
The only place we should depend on `docker/cli` should be the
cli (i.e. `./cmd/docker-app` or some `cli` package if we have that in
the future).

This also use a struct instead of a lenghty list of arguments. The
same should be done with `helm`.
